fix:arrowButton direction 수정
