+++ src/components/ArrowButton/ArrowButton.module.css
@@ -2,7 +2,6 @@
   width: 40px;
   height: 40px;
   border: 1px solid #d9d9d9;
-  border-radius: 4px 0px 0px 4px;
   background-color: #ffffff;
   transition: all 0.25s;
 
@@ -25,4 +24,12 @@
 .arrowButtonImg {
   width: 6.61px;
   height: 12.33px;
+}
+
+.arrowButton.left {
+  border-radius: 4px 0px 0px 4px;
+}
+
+.arrowButton.right {
+  border-radius: 0px 4px 4px 0px;
 }
