--- src/components/ArrowButton/ArrowRightButton.module.css
@@ -1,28 +1,0 @@
-.arrowButton {
-  width: 40px;
-  height: 40px;
-  border: 1px solid #d9d9d9;
-  border-radius: 4px 0px 0px 4px;
-  background-color: #ffffff;
-  transition: all 0.25s;
-
-  &:not(:disabled):hover {
-    border-color: #9fa6b2;
-    background-color: #d9d9d9;
-  }
-
-  &:not(:disabled):active {
-    background-color: #9fa6b2;
-  }
-
-  &:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-    pointer-events: none;
-  }
-}
-
-.arrowButtonImg {
-  width: 6.61px;
-  height: 12.33px;
-}
